:truck: move SignUp component to widgets/sign-up and update imports
